Authetication/Authorization

+++ controllers/clothingItems.js
@@ -1,5 +1,10 @@
 const { ClothingItem } = require("../models/clothingItem");
-const { NOT_FOUND, SERVER_ERROR, BAD_REQUEST } = require("../utils/errors");
+const {
+  NOT_FOUND,
+  SERVER_ERROR,
+  BAD_REQUEST,
+  UNAUTHORIZED,
+} = require("../utils/errors");
 
 // Controller to get all clothing items
 const getItems = async (req, res) => {
@@ -41,21 +46,30 @@ const createItem = async (req, res) => {
 // Controller to delete a clothing item by _id
 const deleteItem = async (req, res) => {
   const { itemId } = req.params;
+  const { _id } = req.user;
   try {
-    const deletedItem = await ClothingItem.findByIdAndRemove(itemId);
-    if (!deletedItem) {
+    const item = await ClothingItem.findById(itemId);
+    if (!item) {
       return res.status(NOT_FOUND).json({ message: "Item not found" });
     }
-    res.json(deletedItem);
+    // Check if the logged-in user is the owner of the item
+    if (item.owner.toString() !== _id) {
+      // If the owner's _id is not the same as the logged-in user's _id, return a 403 error
+      return res
+        .status(UNAUTHORIZED)
+        .json({ message: "You are not authorized to delete this item" });
+    }
+    // Delete the item from the database and return response
+    const deletedItem = await ClothingItem.findByIdAndRemove(itemId);
+    return res.json(deletedItem);
   } catch (err) {
     if (err.name === "CastError") {
       return res.status(BAD_REQUEST).json({ message: "Invalid item ID" });
     }
-    res
+    return res
       .status(SERVER_ERROR)
       .json({ message: "Failed to delete clothing item" });
   }
-  return deleteItem;
 };
 // Like/Dislike Controllers
 const likeItem = (req, res) => {
